+++ exp_2.py
@@ -37,13 +37,6 @@ label_names = ["TEFF", "LOGG", "FE_H", "C_FE", "CI_FE", "N_FE", "O_FE", "NA_FE",
 label_names_for_balancing = ["TEFF", "LOGG", "FE_H", "NA_FE", "O_FE", "MG_FE",
                              "AL_FE"]
 
-#vacuum_wavelengths, training_set_labels, training_set_flux, training_set_ivar \
-#    = get_balanced_training_set(label_names, label_names_for_balancing)
-with open("experiments/2/training_set.pkl", "rb") as fp:
-    vacuum_wavelengths, training_set_labels, training_set_flux, training_set_ivar \
-        = pickle.load(fp)
-
-
 # ----------- #
 train_kwds = dict(op_kwds=dict(factr=1e12, pgtol=1e-5))
 test_kwds = dict()
@@ -52,31 +45,42 @@ OUTPUT_PATH = "experiments/2/"
 
 if not os.path.exists(OUTPUT_PATH): os.mkdir(OUTPUT_PATH)
 
+model_path = os.path.join(OUTPUT_PATH, "restricted.model")
+if os.path.exists(model_path):
+    model = tc.CannonModel.read(model_path)
 
-vectorizer = tc.vectorizer.PolynomialVectorizer(label_names, order=2)
+else:
 
-model = tc.restricted.RestrictedCannonModel(
-    training_set_labels, training_set_flux, training_set_ivar, vectorizer,
-    dispersion=vacuum_wavelengths,
-    theta_bounds=dict([(ln, (None, 0)) for ln in label_names if ln.endswith("_FE")]))
+    #vacuum_wavelengths, training_set_labels, training_set_flux, training_set_ivar \
+    #    = get_balanced_training_set(label_names, label_names_for_balancing)
+    with open("experiments/2/training_set.pkl", "rb") as fp:
+        vacuum_wavelengths, training_set_labels, training_set_flux, training_set_ivar \
+            = pickle.load(fp)
 
-model.train(**train_kwds)
-model.write(os.path.join(OUTPUT_PATH, "restricted.model"), overwrite=True)
+    vectorizer = tc.vectorizer.PolynomialVectorizer(label_names, order=2)
 
-# Plot theta.
-fig = tc.plot.theta(model, indices=np.arange(len(label_names) + 1))
-fig.set_figheight(20)
-fig.tight_layout()
-fig.subplots_adjust(hspace=0, wspace=0)
-fig.savefig(os.path.join(OUTPUT_PATH, "restricted_theta.pdf"), dpi=300)
+    model = tc.restricted.RestrictedCannonModel(
+        training_set_labels, training_set_flux, training_set_ivar, vectorizer,
+        dispersion=vacuum_wavelengths,
+        theta_bounds=dict([(ln, (None, 0)) for ln in label_names if ln.endswith("_FE")]))
+
+    model.train(**train_kwds)
+    model.write(model_path, overwrite=True)
+
+    # Plot theta.
+    fig = tc.plot.theta(model, indices=np.arange(len(label_names) + 1))
+    fig.set_figheight(20)
+    fig.tight_layout()
+    fig.subplots_adjust(hspace=0, wspace=0)
+    fig.savefig(os.path.join(OUTPUT_PATH, "restricted_theta.pdf"), dpi=300)
 
 
-# Do one-to-one.
-oto_labels, oto_cov, oto_meta = model.test(training_set_flux, training_set_ivar,
-                                           **test_kwds)
+    # Do one-to-one.
+    oto_labels, oto_cov, oto_meta = model.test(training_set_flux, training_set_ivar,
+                                               **test_kwds)
 
-fig = tc.plot.one_to_one(model, oto_labels)
-fig.savefig(os.path.join(OUTPUT_PATH, "restricted_one_to_one.pdf"))
+    fig = tc.plot.one_to_one(model, oto_labels)
+    fig.savefig(os.path.join(OUTPUT_PATH, "restricted_one_to_one.pdf"))
 
 
 # Run the model on all visits and plot dispersion as a function of S/N value.
